backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
openai==0.28.1
python-dotenv==1.0.0
pydantic>=2.7,<3.0
python-multipart==0.0.6
